# docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "\"Een API om referentielijstwaarden en de gemeentelijke selectielijst\ntebenaderen. Selectielijst De [Gemeentelijke Selectielijst](https://vng.nl/selectielijst)\nis relevantin het kader van archivering.**Zaakgericht werken** Bij\nhet configureren van zaaktypes (en resultaattypes) in de catalogus APIrefereren\neen aantal resources naar resources binnen de Selectielijst API. Het gaat dan\nom de `ProcesType` en `Resultaat` resources.## ReferentielijstenReferentielijsten\nbevat een standaardset aan waarden. Deze waarden zijn net t\\xE9dynamisch om\nin een enum opgenomen te worden, maar er is wel behoefte om deze landelijk\nte standaardiseren. Een voorbeeld hiervan is de set aan mogelijkecommunicatiekanalen.\n## Autorisatie: Deze APIs zijn alleen-lezen, en behoeven geen autorisatie.\n## Inhoud: De inhoud wordt beheerd door VNG Realisatie. Om de inhoud van\nreferentielijsten bij te werken, contacteer dan VNG Realisatie via e-mail of\nop Github.De inhoud van de Gemeentelijke Selectielijst wordt ge\\xEFmporteerd\nvanuit degepubliceerde Excel-bestanden.\"",
    "title": "Referentielijsten \u0026 Selectielijst API",
    "contact": {
      "url": "https://github.com/VNG-Realisatie/referentielijsten-api",
      "email": "standaarden.ondersteuning@vng.nl"
    },
    "license": {
      "name": "EUPL 1.2",
...
  "host": "referentielijsten-api.internal",
  "basePath": "/api/v1",
  "paths": {
...
    "/procestypen": {
      "get": {
        "description": "Procestypen worden gerefereerd in zaaktypecatalogi - bij het configureren\nvan een zaaktype wordt aangegeven welk procestype van toepassing is, zodat\nhet archiefregime van zaken bepaald kan worden.\nZie https://vng.nl/files/vng/20170706-selectielijst-gemeenten-intergemeentelijke-organen-2017.pdf\nvoor de bron van de inhoud.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "procestypen"
        ],
        "summary": "Ontsluit de selectielijst procestypen.",
        "operationId": "procestypen_list",
        "parameters": [
          {
            "type": "string",
            "description": "jaar",
            "name": "jaar",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "ProcesType",
            "schema": {
              "$ref": "#/definitions/ProcesType"
            }
          },
          "400": {
            "description": "ValidatieFout",
            "schema": {
              "$ref": "#/definitions/ValidatieFout"
            }
          },
          "405": {
            "description": "Fout",
            "schema": {
              "$ref": "#/definitions/Fout"
            }
          }
        }
      }
    },
    "/procestypen/{uuid}": {
      "get": {
        "description": "Procestypen worden gerefereerd in zaaktypecatalogi - bij het configureren\nvan een zaaktype wordt aangegeven welk procestype van toepassing is, zodat\nhet archiefregime van zaken bepaald kan worden.\nZie https://vng.nl/files/vng/20170706-selectielijst-gemeenten-intergemeentelijke-organen-2017.pdf\nvoor de bron van de inhoud.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "procestypen"
        ],
        "summary": "Ontsluit de selectielijst procestypen.",
        "operationId": "procestypen_retrieve",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "uuid van de resultaattypeomschrijving",
            "name": "uuid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ProcesType",
            "schema": {
              "$ref": "#/definitions/ProcesType"
            }
          },
          "400": {
            "description": "ValidatieFout",
            "schema": {
              "$ref": "#/definitions/ValidatieFout"
            }
          },
          "404": {
            "description": "Fout",
            "schema": {
              "$ref": "#/definitions/Fout"
            }
          },
          "405": {
            "description": "Fout",
            "schema": {
              "$ref": "#/definitions/Fout"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "FieldValidationError": {
      "description": "Formaat van validatiefouten.",
      "type": "object",
      "title": "FieldValidationError represents an error",
      "required": [
        "name",
        "code",
        "reason"
      ],
      "properties": {
        "code": {
          "description": "Systeemcode die het type fout aangeeft",
          "type": "string"
        },
        "name": {
          "description": "Naam van het veld met ongeldige gegevens",
          "type": "string"
        },
        "reason": {
          "description": "Uitleg wat er precies fout is met de gegevens",
          "type": "string"
        }
      }
    },
    "Fout": {
      "description": "Formaat van HTTP 4xx en 5xx fouten.",
      "type": "object",
      "title": "Fout represents an error",
      "required": [
        "code",
        "title",
        "status",
        "detail",
        "instance"
      ],
      "properties": {
        "code": {
          "description": "Systeemcode die het type fout aangeeft",
          "type": "string"
        },
        "detail": {
          "description": "Extra informatie bij de fout, indien beschikbaar",
          "type": "string"
        },
        "instance": {
          "description": "URI met referentie naar dit specifiek voorkomen van de fout. Deze kan gebruikt worden in combinatie met server logs, bijvoorbeeld.",
          "type": "string"
        },
        "status": {
          "description": "De HTTP status code",
          "type": "integer",
          "format": "int64"
        },
        "title": {
          "description": "Generieke titel voor het type fout",
          "type": "string"
        },
        "type": {
          "description": "URI referentie naar het type fout, bedoeld voor developers",
          "type": "string"
        }
      }
    },
...
    "ProcesType": {
      "type": "object",
      "required": [
        "url",
        "nummer",
        "jaar",
        "naam",
        "omschrijving",
        "toelichting",
        "procesobject"
      ],
      "properties": {
        "jaar": {
          "description": "Nummer van de selectielijstcategorie",
          "type": "integer",
          "format": "int32",
          "maximum": 2147483647,
          "minimum": -2147483648,
          "example": 2020
        },
        "naam": {
          "description": "Benaming van het procestype",
          "type": "string",
          "maxLength": 100,
          "example": "Producten en diensten leveren"
        },
        "nummer": {
          "description": "Het jaartal waartoe dit ProcesType behoort",
          "type": "integer",
          "format": "int32",
          "maximum": 2147483647,
          "minimum": -2147483648,
          "example": 6
        },
        "omschrijving": {
          "description": "Omschrijving van het procestype",
          "type": "string",
          "maxLength": 300,
          "example": "Het door het orgaan leveren van (publieke) producten of diensten"
        },
        "procesobject": {
          "description": "Object waar de uitvoering van het proces op van toepassing\nis en waarvan de bestaans- of geldigheidsduur eventueel van belang is\nbij het bepalen van de start van de bewaartermijn",
          "type": "string",
          "maxLength": 80,
          "example": "Het geleverde product of de geleverde dienst"
        },
        "toelichting": {
          "description": "Toelichting van het procestype",
          "type": "string",
          "example": "Voor het verstrekken van een product of dienst hoeft de aanvrager niet aan een bepaalde situatie te voldoen, want in dat geval is het procestype 'Voorzieningen verstrekken' van toepassing. Het product of de dienst houdt niet het doen of laten van het orgaan in, want in dat geval valt het onder het procestype 'Verzoeken behandelen'."
        },
        "url": {
          "description": "Interne referentie naar de url",
          "type": "string",
          "readOnly": true,
          "example": "http://referentielijsten-api.internal/api/v1/651a1b5b-f84f-4c73-9151-4d485c7dcb99"
        }
      }
    },
...
    "ValidatieFout": {
      "description": "Formaat van HTTP 4xx en 5xx fouten.",
      "type": "object",
      "title": "ValidatieFout represents an error",
      "required": [
        "code",
        "title",
        "status",
        "detail",
        "instance",
        "invalidParams"
      ],
      "properties": {
        "code": {
          "description": "Systeemcode die het type fout aangeeft",
          "type": "string"
        },
        "detail": {
          "description": "Extra informatie bij de fout, indien beschikbaar",
          "type": "string"
        },
        "instance": {
          "description": "URI met referentie naar dit specifiek voorkomen van de fout. Deze kan gebruikt worden in combinatie met server logs, bijvoorbeeld.",
          "type": "string"
        },
        "invalidParams": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FieldValidationError"
          }
        },
        "status": {
          "description": "De HTTP status code",
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "description": "Generieke titel voor het type fout",
          "type": "string"
        },
        "type": {
          "description": "URI referentie naar het type fout, bedoeld voor developers",
          "type": "string"
        }
      }
    }
  }
}